login funcionando

+++ index.js
@@ -1,7 +1,14 @@
 const express = require('express');
+
+const CryptoJS = require("crypto-js");
+const jwt = require('jsonwebtoken');
+var SHA256 = require("crypto-js/sha256");
+var MD5 = require("crypto-js/md5");
+
 const server = express();
 server.use(express.json());
 const db = require("./db");
+
 
 function isPhoneCorrect(phone) {
     const splittedPhone = phone.split('');
@@ -22,18 +29,18 @@ function isPhoneCorrect(phone) {
 }
 
 server.get('/contato/:id', (req, res) => {
-    const {id} = req.params; 
+    const {id} = req.params;     
     (async () => {
         try {
             const conn = await db.connect();
             const [rows] = await conn.query(`select nome, celular from contacts where id = ${id};`);
-            
             if(rows.length < 1) res.json({message: "Nenhum contato encontrato"});
             
             const resultArray = Object.values(JSON.parse(JSON.stringify(rows)))
-            return res.json(resultArray[0]);
+            return res.status(200).json(resultArray[0]);
         } catch (error) {
-            return res.json({message: "Ocorreu um erro ao buscar contato"});
+            console.log(error);
+            res.status(400).json({message:  "Ocorreu um erro ao buscar contato"});
         }
     })();
 })
@@ -47,9 +54,10 @@ server.get('/contatos', (req, res) => {
             if(rows.length < 1) res.json({message: "Nenhum contato encontrato"});
 
             const resultArray = Object.values(JSON.parse(JSON.stringify(rows)))
-            return res.json(resultArray);
+            return res.status(200).json(resultArray);
         } catch (error) {
-            return res.json({message: "Ocorreu um erro ao buscar contatos"});
+            console.log(error);
+            res.status(400).json({message:  "Ocorreu um erro ao buscar contatos"});
         }
     })();
 })
@@ -67,10 +75,11 @@ server.post('/contato', (req, res) => {
                 
                 const conn = await db.connect();
                 await conn.query(`insert into contacts (nome, celular) values ('${contato.name}', '${contato.phone}');`); 
-                return res.json({message: `Contatos cadastrados com sucesso`});
+                return res.status(200).json({message: `Contatos cadastrados com sucesso`});
             })
         } catch (error) {
-            return res.json({message: "Ocorreu um erro ao cadastrar contatos"});
+            console.log(error);
+            res.status(400).json({message: "Ocorreu um erro ao cadastrar contatos"});
         }
     })();
 })
@@ -85,10 +94,15 @@ server.put('/contato/:id', (req, res) => {
             return res.json({message: 'Somente letras maiusculas devem ser utilizadas no nome'})
         
             const conn = await db.connect();
-             await conn.query(`UPDATE contacts SET nome = '${name}', celular = '${phone}' WHERE id=${id};`); 
-             return res.json({message: "Usuário atualizado com sucesso"});
+            
+            const [rows] = await conn.query(`select * from contacts where id=${id};`);
+            if(rows.length < 1) return res.json({message: "Usuário não encontrado"});
+
+            await conn.query(`UPDATE contacts SET nome = '${name}', celular = '${phone}' WHERE id=${id};`); 
+            return res.status(200).json({message: "Usuário atualizado com sucesso"});
             } catch (error) {
-            return res.json({message: "Ocorreu um erro ao atualizar contatos"});
+                console.log(error);
+                res.status(400).json({message:  "Ocorreu um erro ao atualizar contatos"});
         }
     })();
 })
@@ -98,12 +112,42 @@ server.delete('/contato/:id', (req, res) => {
     (async () => {
         try {            
             const conn = await db.connect();
+            const [rows] = await conn.query(`select * from contacts where id=${id};`);
+            if(rows.length < 1) return res.json({message: "Usuário não encontrado"});
+    
             await conn.query(`delete from contacts where id=${id};`); 
-            return res.json({message: "Usuário deletado com sucesso"});
+            return res.status(200).json({message: "Usuário deletado com sucesso"});
         } catch (error) {
-            return res.json({message: "Ocorreu um erro ao deletar contatos"});
+            console.log(error);
+            res.status(400).json({message: "Ocorreu um erro ao deletar contatos"});
         }
     })();
 })
+
+server.post('/auth', (req, res) => {
+    const {email, password} = req.body;
+    (async () => {
+        try {
+            const conn = await db.connect();
+            const [rows] = await conn.query(`select * from users where email='${email}';`);
+
+            if(rows.length < 1) return res.json({message: "Usuário não encontrado"});
+            const users = Object.values(JSON.parse(JSON.stringify(rows)))
+
+            const encryptedPassword = CryptoJS.MD5(password);
+            if (users[0].senha !== encryptedPassword.toString()) return res.status(401).json({message: "Ocorreu um erro na autenticação"});
+
+            const authenticatedUser = users[0];
+            const token = jwt.sign({ authenticatedUser }, 'SECRET', {
+              expiresIn: "7d"
+            });
+            return res.status(200).json(token);
+           } catch (error) {
+            console.error(error)
+            return res.status(401).json({message: "Ocorreu um erro na autenticação"});
+          }
+    })();
+})
+
 
 server.listen(3000);
